+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.1.0
+
+- Password recovery & email settings
+- Autofocus guess input on play page
+- Fix "try another" button when 1 screenshot left to be found
+
 # v1.0.0
 
 - Login & registration